refactor: centralize skill setup in app_conversation_service_base and rename git_app_conversation_service accordingly
