+++ frontend/src/i18n/declaration.ts
@@ -440,6 +440,7 @@ export enum I18nKey {
   STATUS$STARTING_RUNTIME = "STATUS$STARTING_RUNTIME",
   STATUS$SETTING_UP_WORKSPACE = "STATUS$SETTING_UP_WORKSPACE",
   STATUS$SETTING_UP_GIT_HOOKS = "STATUS$SETTING_UP_GIT_HOOKS",
+  STATUS$SETTING_UP_SKILLS = "STATUS$SETTING_UP_SKILLS",
   ACCOUNT_SETTINGS_MODAL$DISCONNECT = "ACCOUNT_SETTINGS_MODAL$DISCONNECT",
   ACCOUNT_SETTINGS_MODAL$SAVE = "ACCOUNT_SETTINGS_MODAL$SAVE",
   ACCOUNT_SETTINGS_MODAL$CLOSE = "ACCOUNT_SETTINGS_MODAL$CLOSE",
