+++ frontend/src/types/runtime-status.ts
@@ -5,6 +5,7 @@ export type RuntimeStatus =
   | "STATUS$RUNTIME_STARTED"
   | "STATUS$SETTING_UP_WORKSPACE"
   | "STATUS$SETTING_UP_GIT_HOOKS"
+  | "STATUS$SETTING_UP_SKILLS"
   | "STATUS$READY"
   | "STATUS$ERROR"
   | "STATUS$ERROR_RUNTIME_DISCONNECTED"
