+++ openhands/app_server/app_conversation/app_conversation_models.py
@@ -118,6 +118,7 @@ class AppConversationStartTaskStatus(Enum):
     PREPARING_REPOSITORY = 'PREPARING_REPOSITORY'
     RUNNING_SETUP_SCRIPT = 'RUNNING_SETUP_SCRIPT'
     SETTING_UP_GIT_HOOKS = 'SETTING_UP_GIT_HOOKS'
+    SETTING_UP_SKILLS = 'SETTING_UP_SKILLS'
     STARTING_CONVERSATION = 'STARTING_CONVERSATION'
     READY = 'READY'
     ERROR = 'ERROR'
